+++ app/views/personaldetails/addressSelector.scala.html
@@ -60,7 +60,7 @@ partialRetriever: LocalPartialRetriever, templateRenderer: TemplateRenderer, mes
             }
         </fieldset>
 
-        <p class="form-group"><a href="@controllers.routes.AddressController.showUpdateAddressForm(typ)">@messages("label.enter_your_address_if_it_is_not_shown")</a></p>
+        <p class="form-group"><a href="@controllers.address.routes.UpdateAddressController.onPageLoad(typ)">@messages("label.enter_your_address_if_it_is_not_shown")</a></p>
 
         <input type="submit" class="button" id="submitAddressSelector" value='@messages("global.label.continue_button")' >
       }
