+++ app/views/personaldetails/updateAddress.scala.html
@@ -52,7 +52,7 @@ configDecorator: ConfigDecorator, partialRetriever: LocalPartialRetriever, templ
         </h1>
       </header>
 
-      @helpers.form(action = controllers.routes.AddressController.processUpdateAddressForm(typ), args = 'class -> "update-address-form", 'id -> "update-address-form") {
+      @helpers.form(action = controllers.address.routes.UpdateAddressController.onSubmit(typ), args = 'class -> "update-address-form", 'id -> "update-address-form") {
 
         <div class="form-group">
           @input(
