+++ conf/prod.routes
@@ -47,8 +47,8 @@ GET         /personal-account/your-address/:typ/find-address                    
 POST        /personal-account/your-address/:typ/find-address                           controllers.address.PostcodeLookupController.onSubmit(typ: AddrType)
 GET         /personal-account/your-address/:typ/select-address                         controllers.address.AddressSelectorController.onPageLoad(typ: AddrType)
 POST        /personal-account/your-address/:typ/select-address                         controllers.address.AddressSelectorController.onSubmit(typ: AddrType)
-GET         /personal-account/your-address/:typ/edit-address                           controllers.AddressController.showUpdateAddressForm(typ: AddrType)
-POST        /personal-account/your-address/:typ/edit-address                           controllers.AddressController.processUpdateAddressForm(typ: AddrType)
+GET         /personal-account/your-address/:typ/edit-address                           controllers.address.UpdateAddressController.onPageLoad(typ: AddrType)
+POST        /personal-account/your-address/:typ/edit-address                           controllers.address.UpdateAddressController.onSubmit(typ: AddrType)
 GET         /personal-account/your-address/:typ/enter-international-address            controllers.AddressController.showUpdateInternationalAddressForm(typ: AddrType)
 POST        /personal-account/your-address/:typ/enter-international-address            controllers.AddressController.processUpdateInternationalAddressForm(typ: AddrType)
 GET         /personal-account/your-address/:typ/enter-start-date                       controllers.AddressController.enterStartDate(typ: AddrType)
