+++ test/controllers/AddressControllerSpec.scala
@@ -166,286 +166,6 @@ class AddressControllerSpec extends BaseSpec with MockitoSugar {
     })
   }
 
-  "Calling AddressController.showUpdateAddressForm" should {
-
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-    }
-
-    "find only the selected address from the session cache and no residency choice and return 303" in new LocalSetup {
-
-      lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("soleSelectedAddressRecord" -> Json.toJson(fakeStreetPafAddressRecord))))
-
-      val result = controller.showUpdateAddressForm(SoleAddrType)(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/personal-details")
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "fetch the selected address and a sole residencyChoice has been selected from the session cache and return 200" in new LocalSetup {
-
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "selectedAddressRecord" -> Json.toJson(fakeStreetPafAddressRecord),
-            "soleResidencyChoiceDto" -> Json.toJson(ResidencyChoiceDto(PostalAddrType)))))
-
-      val result = controller.showUpdateAddressForm(SoleAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "find no selected address with sole address type but residencyChoice in the session cache and still return 200" in new LocalSetup {
-
-      lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("soleResidencyChoiceDto" -> Json.toJson(ResidencyChoiceDto(SoleAddrType)))))
-
-      val result = controller.showUpdateAddressForm(SoleAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "find no residency choice in the session cache and redirect to the beginning of the journey" in new LocalSetup {
-
-      lazy val sessionCacheResponse = Some(CacheMap("id", Map.empty))
-
-      val result = controller.showUpdateAddressForm(SoleAddrType)(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      redirectLocation(result) shouldBe Some("/personal-account/personal-details")
-    }
-
-    "redirect user to beginning of journey and return 303 for postal addressType and no pagevisitedDto in cache" in new LocalSetup {
-
-      lazy val sessionCacheResponse = None
-
-      val result = controller.showUpdateAddressForm(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      redirectLocation(result) shouldBe Some("/personal-account/personal-details")
-    }
-
-    "display edit address page and return 200 for postal addressType with pagevisitedDto and addressRecord in cache" in new LocalSetup {
-
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "addressPageVisitedDto" -> Json.toJson(AddressPageVisitedDto(true)),
-            "selectedAddressRecord" -> Json.toJson(fakeStreetPafAddressRecord))))
-
-      val result = controller.showUpdateAddressForm(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "display edit address page and return 200 for postal addressType with pagevisitedDto and no addressRecord in cache" in new LocalSetup {
-
-      lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("addressPageVisitedDto" -> Json.toJson(AddressPageVisitedDto(true)))))
-
-      val result = controller.showUpdateAddressForm(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "find no addresses in the session cache and return 303" in new LocalSetup {
-
-      lazy val sessionCacheResponse = Some(CacheMap("id", Map.empty))
-
-      val result = controller.showUpdateAddressForm(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/personal-details")
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "find sole selected and submitted addresses in the session cache and return 200" in new LocalSetup {
-
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "soleSelectedAddressRecord" -> Json.toJson(fakeStreetPafAddressRecord),
-            "soleSubmittedAddressDto" -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified)),
-            "soleResidencyChoiceDto" -> Json.toJson(ResidencyChoiceDto(SoleAddrType))
-          )
-        ))
-
-      val result = controller.showUpdateAddressForm(SoleAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "find no selected address but a submitted address in the session cache and return 200" in new LocalSetup {
-
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "soleSubmittedAddressDto" -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified)),
-            "soleResidencyChoiceDto" -> Json.toJson(ResidencyChoiceDto(SoleAddrType))
-          )
-        ))
-
-      val result = controller.showUpdateAddressForm(SoleAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "show 'Enter the address' when user amends correspondence address manually and address has not been selected" in new LocalSetup {
-
-      lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("addressPageVisitedDto" -> Json.toJson(AddressPageVisitedDto(true)))))
-
-      val result = controller.showUpdateAddressForm(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      val doc = Jsoup.parse(contentAsString(result))
-      doc.getElementsByClass("heading-xlarge").toString().contains("Your postal address") shouldBe true
-    }
-
-    "show 'Enter your address' when user amends residential address manually and address has not been selected" in new LocalSetup {
-
-      lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("soleResidencyChoiceDto" -> Json.toJson(ResidencyChoiceDto(SoleAddrType)))))
-
-      val result = controller.showUpdateAddressForm(SoleAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      val doc = Jsoup.parse(contentAsString(result))
-      doc.getElementsByClass("heading-xlarge").toString().contains("Your address") shouldBe true
-    }
-
-    "show 'Edit the address (optional)' when user amends correspondence address manually and address has been selected" in new LocalSetup {
-
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "addressPageVisitedDto" -> Json.toJson(AddressPageVisitedDto(true)),
-            "postalSelectedAddressRecord" -> Json.toJson(Fixtures.fakeStreetPafAddressRecord)
-          )
-        ))
-
-      val result = controller.showUpdateAddressForm(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      val doc = Jsoup.parse(contentAsString(result))
-      doc.getElementsByClass("heading-xlarge").toString().contains("Edit the address (optional)") shouldBe true
-    }
-
-    "show 'Edit your address (optional)' when user amends residential address manually and address has been selected" in new LocalSetup {
-
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "soleResidencyChoiceDto" -> Json.toJson(ResidencyChoiceDto(SoleAddrType)),
-            "soleSelectedAddressRecord" -> Json.toJson(Fixtures.fakeStreetPafAddressRecord)
-          )
-        ))
-
-      val result = controller.showUpdateAddressForm(SoleAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      val doc = Jsoup.parse(contentAsString(result))
-      doc.getElementsByClass("heading-xlarge").toString().contains("Edit your address (optional)") shouldBe true
-    }
-  }
-
-  "Calling AddressController.processUpdateAddressForm" should {
-
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("addressLookupServiceDown" -> Json.toJson(Some(true)))))
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-    }
-
-    "return 400 when supplied invalid form input" in new LocalSetup {
-
-      override lazy val sessionCacheResponse = Some(CacheMap("id", Map("selectedAddressRecord" -> Json.toJson(""))))
-
-      val result = controller.processUpdateAddressForm(PostalAddrType)(FakeRequest("POST", ""))
-
-      status(result) shouldBe BAD_REQUEST
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "return 303, caching addressDto and redirecting to enter start date page when supplied valid form input on a postal journey" in new LocalSetup {
-
-      val form = FakeRequest("POST", "").withFormUrlEncodedBody(fakeStreetTupleListAddressForUnmodified: _*)
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = form
-            )
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.processUpdateAddressForm(PostalAddrType)(form)
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/your-address/postal/changes")
-      verify(controller.sessionCache, times(1)).cache(
-        meq("postalSubmittedAddressDto"),
-        meq(asAddressDto(fakeStreetTupleListAddressForUnmodified)))(any(), any(), any())
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "return 303, caching addressDto and redirecting to review changes page when supplied valid form input on a non postal journey and input default startDate into cache" in new LocalSetup {
-
-      val form = FakeRequest("POST", "").withFormUrlEncodedBody(fakeStreetTupleListAddressForUnmodified: _*)
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture{
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = form
-            )
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.processUpdateAddressForm(SoleAddrType)(form)
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/your-address/sole/changes")
-      verify(controller.sessionCache, times(1)).cache(
-        meq("soleSubmittedAddressDto"),
-        meq(asAddressDto(fakeStreetTupleListAddressForUnmodified)))(any(), any(), any())
-      verify(controller.sessionCache, times(1))
-        .cache(meq("soleSubmittedStartDateDto"), meq(DateDto(LocalDate.now())))(any(), any(), any())
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-  }
-
   "Calling AddressController.showUpdateInternationalAddressForm" should {
 
     trait LocalSetup extends WithAddressControllerSpecSetup {
